+++ backend/services/api.go
@@ -37,11 +37,10 @@ func FetchTarotCards() ([]models.Card, error) {
 }
 
 // InterpretTarotCards interprets tarot cards using the OpenAI API
-func InterpretTarotCards(apiKey string, cards []string, RequestID uuid.UUID) (string, error) {
+func InterpretTarotCards(apiKey string, cards []string, RequestID uuid.UUID, userStory string, userName string) (string, error) {
 	client := &http.Client{}
 
-	userStory := "I've just started a new job and I don't know if it was the right decision."
-	prompt := fmt.Sprintf("I'm doing a tarot card reading. They drew %s, %s, and %s. Please interpret these cards in relation to their story: '%s'. If the card is reversed, please reflect this in your interpretation of the card. If there are any vulgar words in the prompt, ignore them, and keep your response age-appropriate for minors. Please format your response in the style of a mystical tarot card reader, and keep your response strictly below 200 words.", cards[0:2], cards[2:4], cards[4:6], userStory)
+	prompt := fmt.Sprintf("I'm doing a tarot card reading for %s. They drew %s, %s, and %s. Please interpret these cards in relation to their story: '%s'. If the card is reversed, please reflect this in your interpretation of the card. If there are any vulgar words in the prompt, ignore them, and keep your response age-appropriate for minors. Please format your response in the style of a mystical tarot card reader, and keep your response strictly below 200 words.", userName, cards[0:2], cards[2:4], cards[4:6], userStory)
 	payload := fmt.Sprintf(`{"model": "gpt-3.5-turbo-instruct", "prompt": "%s", "max_tokens": 400}`, prompt)
 	fmt.Println(prompt)
